fix(webui): spotify credentials not saving
