+++ deezer-sdk/tsconfig.json
@@ -4,7 +4,7 @@
 		"rootDir": "./src",
 		"outDir": "./dist",
 		"strict": false,
-		"strictNullChecks": true,
+		// "strictNullChecks": true,
 		"types": ["vitest/globals"]
 	},
 	"include": ["./src/**/*.ts"]
