+++ webui/src/server/websocket/modules/saveSettings.ts
@@ -1,7 +1,7 @@
 import { WebSocketServer } from "ws";
 import { logger } from "@/helpers/logger.js";
 import { DeemixApp } from "@/deemixApp.js";
-import type { Settings, SpotifySettings } from "@/types.js";
+import type { Settings, SpotifySettings } from "deemix";
 
 const eventName = "saveSettings";
